refactor: get balance
